Refactor the reading logic as the API is inconsitent.
We can't rely on EOF as some commands return more than a single line.
For example  with  `(l *Lora) Send(data string)` if you rely on EOF
sometime is till return OK and sometimes it will return OKat+recv=2,0,0
(the second part is the ack package). That is way it is important to
know exactly how many lines want to read from the output.
I spoke with one of the hat developers and he confirmed that they are
doing the same thing - reading a given number of lines based on the
command.
The PR also adds a global timeout as  for the serial library has a max
timeout of 25 seconds even when opening the port with a higher timeout
in the cofig settings.

Signed-off-by: Krasi Georgiev <8903888+krasi-georgiev@users.noreply.github.com>